move GetPathContent to test_data_util
